Move first unittest to shared location.
