+++ ush/generate_FV3LAM_wflow.py
@@ -788,52 +788,3 @@ if __name__ == "__main__":
     )
 
 
-class Testing(unittest.TestCase):
-    def test_generate_FV3LAM_wflow(self):
-
-        # run workflows in separate process to avoid conflict between community and nco settings
-        def run_workflow(USHdir, logfile):
-            p = Process(target=generate_FV3LAM_wflow, args=(USHdir, logfile))
-            p.start()
-            p.join()
-            exit_code = p.exitcode
-            if exit_code != 0:
-                sys.exit(exit_code)
-
-        USHdir = os.path.dirname(os.path.abspath(__file__))
-        logfile = "log.generate_FV3LAM_wflow"
-        SED = get_env_var("SED")
-
-        # community test case
-        cp_vrfy(f"{USHdir}/config.community.yaml", f"{USHdir}/config.yaml")
-        run_command(
-            f"""{SED} -i 's/MACHINE: hera/MACHINE: linux/g' {USHdir}/config.yaml"""
-        )
-        run_workflow(USHdir, logfile)
-
-        # nco test case
-        nco_test_config = load_config_file(f"{USHdir}/config.nco.yaml")
-        # Since we don't have a pre-gen grid dir on a generic linux
-        # platform, turn the make_* tasks on for this test.
-        cfg_updates = {
-            "user": {
-                "MACHINE": "linux",
-            },
-            "rocoto": {
-                "tasks": {
-                    "taskgroups": \
-                        '\'{{ ["parm/wflow/prep.yaml","parm/wflow/coldstart.yaml", "parm/wflow/post.yaml"]|include }}\''
-                },
-            },
-        }
-        update_dict(cfg_updates, nco_test_config)
-
-        with open(f"{USHdir}/config.yaml", "w") as cfg_file:
-            cfg_file.write(cfg_to_yaml_str(nco_test_config))
-
-        run_workflow(USHdir, logfile)
-
-    def setUp(self):
-        define_macos_utilities()
-        set_env_var("DEBUG", False)
-        set_env_var("VERBOSE", False)
